[UI] prevent paying off partial loan with more cash than player's cash
At https://18xx.games/game/190389?action=1060, a player that had $605 cash has
just paid $405 off of their 1134 loan. The next action is another partial pay
off for $405--because the input still read 405 from their previous action--but
their cash total is now $200. The engine correctly processes a payment of $200,
despite the action having `@amount = 405`.

With this change, the input's value is updated to not exceed the player's cash
total.